Fix linux game bug
