+++ pages/api/css.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 
 export default async function css(req, res) {
-  const data = fs.readFileSync('games/css.js');
+  const data = fs.readFileSync('games/css/main.js');
   res.status(200).json(data);
 }
