+++ pages/linux.js
@@ -6,9 +6,20 @@ import styles from '../styles/modules/game.module.scss';
 import Script from 'next/script';
 import fs from 'fs';
 
-export default function Game({ game }) {
+export default function Linux({ game }) {
   return (
     <Layout>
+      <div
+        id="game_container"
+        style={{
+          position: 'absolute',
+          left: '360px',
+          top: '82.1px',
+          width: '720px',
+          height: '672.8px',
+          zIndex: 1
+        }}
+      />
       <div className={styles.container}>
         <div className={styles.header}>
           <Nav />
@@ -22,7 +33,7 @@ export default function Game({ game }) {
 };
 
 export async function getServerSideProps(context) {
-  const game = fs.readFileSync('games/linux.js', { encoding: 'utf8' });
+  const game = fs.readFileSync('games/linux/main.js', { encoding: 'utf8' });
   return {
     props: { game }
   };
